Split the arch building and pushing separately
When building multi-arch it's done in parallel which can result
in the same layer being pushed at the same time, which is bad
for the registries.

Instead, build everything in paralle, then push each arch independently,
then push all the arches again to cause the manifest to be written
and pushed properly.

Change-Id: I8036a9b4d4c515c20a05994741540b999e7cbcae