Run sphinx-build in parallel for releasenotes
Since the recent releases and upper-constraints bumps for
openstackdocstheme 2.1.0 and reno 3.1.0, both sphinx extensions now
support safe parallel reading.

This adds the '-j auto' option to the sphinx-build command for building
release notes to parallelize the build over multiple CPUs.

This effort was inspired by the nova build-openstack-releasenotes job
which typically takes 52m to 1h to run. The job timeout is 1h, so we
are occasionally seeing TIMED_OUT job failures.

With parallel sphinx-build, the job run time is reduced down to 15m.

Change-Id: If3c615119b1a72ce294e22685ddc2dcd6b63f4a0