Simplify twine invocation for PyPI uploads
Modern PyPI no longer requires a separate register step, new
projects are registered at the time of upload, so drop the
registration tasks. Twine can also upload multiple artifacts in a
single invocation, and does so in the order in which they're
provided on the command line, thus there's no longer any need for
looping or different wheel and sdist tarball upload tasks.

Change-Id: I09acaf458af600137c72a70a8782c47b67937b60