ara-report: add option for artifact prefix
After https://review.opendev.org/#/c/742971 the report is generated on
the host.  For system-config, this means the report ends up copied to
"bridge.openstack.org/ara-report" in the logs directory; but the
artifact points to the non-existant top-directory.

Add a variable to allow a top-level dir to be set.

Needed-By: https://review.opendev.org/747101
Change-Id: I61a7d7328b5a45fc8dee09909f9411c397dd31b6