Disable broken fetch-sphinx-tarball test job
Recent security fixes made to zuul made impossible to run this
test. Disabling this test until we can find a way to test it as
this is not an easy change.

This is needed because there were at least two other changes that
failed to pass because triggered this broken job. This was broken
for more than a month.

Main challenge here is that Ansible does not have any `mv` module, so
we still need to use shell. Replacing mv with copy could risk producing
out of disk space failures sooner or later.

Change-Id: If09e0430033fd2f1d6fc5e35e5ec0b10dc7c2dfe
Needed-By: https://review.opendev.org/#/c/748480/
Needed-By: https://review.opendev.org/#/c/748682/