Use Python 3.x with launchpadlib
The add-launchpad-credentials role installs the distro's
python-launchpadlib package which, for the distros carrying that
package at all, sets it up for Python 2.7. At this point Ubuntu has
dropped the python-launchpadlib package entirely from their latest
LTS release and Debian will do the same in their next release, but
they have also carried a python3-launchpadlib package for years.
Switch to it so this will work on newer systems.

Change-Id: I92421cae32bf2da5851144337820685c79ec19ee