Temporarily stop running Gentoo base role tests
The zuul-jobs-test-base-roles-gentoo-17-0-systemd job isn't passing
at the moment because Gentoo now needs Python 3.8 to install
pip/wheel, but the images are still using 3.6. A change has been
merged to correct this, but OpenDev has been unable to build new
Gentoo images since May 2020, so that will need to be solved before
the job will work again. In the meantime, stop running it so our
changes aren't indefinitely blocked. We'll need to apply extra
scrutiny to (or simply hold) any Gentoo-related changes for now
until we're testing that again.

Change-Id: I276bc596c50e7185fae828afa79c73ca9d955ec5