Fix CentOS wheel mirror URL
For CentOS we are currently using default values for "wheel_mirror" from
roles/configure-mirrors/defaults/main.yaml, which means that the URL
uses the full distribution version, which is wrong, we should be using
the major version instead.

As an example, centos-8 jobs are trying to use
  https://mirror.bhs1.ovh.opendev.org/wheel/centos-8.3-x86_64
When they should be using
  https://mirror.bhs1.ovh.opendev.org/wheel/centos-8-x86_64

Which makes some jobs fail to find alembic wheel files and then they
fail to build from source because the CI was expecting the wheel mirror
to have it.

This patch fixes this by creating a Centos specific "wheel_mirror"
variable.

Change-Id: I11199cf943a3070fd62bf486fe2b06381db4b04d
(cherry picked from commit c9cad38eebf1bdabb10b79699598d6ba0f098f54)