Test Debian Buster and Bullseye
Debian Buster is the current stable, and Bullseye is the impending
stable (likely within the month). The parent of this change aims to
fix security mirror configuration for Bullseye, but to do that we
need to add testing for it. While we're at it, drop the Stretch
(oldoldstable) jobs to keep the job count from exploding too much.

Change-Id: I278246eff347dd4bdb3471a8db95d30fcba0f17a