Fix a bug in s3 log uploader with .gz files
When trying to upload .gz files, we do not set the ContentEncoding
argument to upload_fileobj but leave it as None. The upload then fails
because NoneType is not allowed. Therefore, leave out this parameter,
and also the ContentType parameter, from the extra args completely if
they are not set.

Change-Id: I601944ac83d5e823aa4dcfd0db880a38474288af