Remove tumbleweed jobs
OpenDev doesn't currently have an image built for tumblweed. Rather than
try and fix this image up and get it running again OpenDev has proposed
it simply be removed. In preparation for that and to avoid running jobs
for a system that doesn't currently have an image we remove the jobs
here.

It was hoped that tumbleweed would be a good forward looking platform
but it seems to need a lot more care than we are able to give it.
Additionally Zuul is probably the major consumer of the platform and it
doesn't see much consumption by other projects.

Change-Id: Iad351b6da1f08183bc4c49c049c303a9491af844