ensure-podman: add containernetworking-plugins
I think this package is technically optional because you can run as
non-root without it.  But since this is a generic role that is just
supposed to give you a working podman, include it.

Somehow we forgot to remove the podman centos-8 (not stream) job in
previous cleanups; remove it here.

Change-Id: I0a87545d29a5ea49a2d9d98c09467c89e7232d4b