encrypt-file: roll back extended file stat
Change I8e7bc38c68c224795630b90a1b989098a7661491 switched the
"Validate input file" task to use a path stat, but when a list of
files is passed into encrypt_file this fails because the list is
treated as a string. Switch back to the more naive "is undefined"
check until it can be properly redone with a loop or similar and
appropriate tests added.

Change-Id: I83e665bc890aec42462776e0079457bb4506552f