build-docker-image: Fix buildset registry pushes
If we build an image with `docker_registry` set, it will fail to
tag the image for the buildset registry since we do not prefix the
image with `docker_registry` when retagging it (if set).

This patch resolves this by adding that prefix if it is set so
it will refer to an image that exists.

Change-Id: Iba29161156c8e8ff4f79a92771456cf105d780fe